port code to refresh URL params to a new file `useFreshParams.ts`
